create a separate function.
